Puttin' in some werk on getting the file/dir listing.
Try hitting `/TRIGGER/s3files/list` for a JSON dump of files and dirs.

For retrieving a nested dir, use a querystring parameter `uri` which should be the path relative from the field's root. So, for example, to get a list of files and dirs in a directory called `balls` in the root, the request would be formed thusly: `TRIGGER/s3files/list?uri=balls`. For further nested dirs, something like `TRIGGER/s3files/list?uri=cock/and/balls`.